Fix a potential error in duty cycle calculation
Now compute duty cycle only if the conditions are met T2 > T1 and T1 >
T0.

A calculation error would occur if the millisecond timer overflow
during a pump cycle. This only happed even 49 days. In those case we
will assume the duty cycle did not changed. This is a reasonable
assumption that cause little error in the calculation of the pump flow
rate.